moved application load balancer resources to a separate module, following discussion
