+++ troposphere/elasticloadbalancing.py
@@ -75,53 +75,6 @@ class AccessLoggingPolicy(AWSProperty):
     }
 
 
-class LoadBalancerAttributes(AWSProperty):
-    props = {
-        'Key': (basestring, False),
-        'Value': (basestring, False)
-    }
-
-
-class Certificate(AWSProperty):
-    props = {
-        'CertificateArn': (basestring, False)
-    }
-
-
-class Action(AWSProperty):
-    props = {
-        'TargetGroupArn': (basestring, True),
-        'Type': (basestring, True)
-    }
-
-
-class Condition(AWSProperty):
-    props = {
-        'Field': (basestring, True),
-        'Values': ([basestring], True)
-    }
-
-
-class Matcher(AWSProperty):
-    props = {
-        'HttpCode': (basestring, False)
-    }
-
-
-class TargetGroupAttribute(AWSProperty):
-    props = {
-        'Key': (basestring, False),
-        'Value': (basestring, False)
-    }
-
-
-class TargetDescription(AWSProperty):
-    props = {
-        'Id': (basestring, True),
-        'Port': (network_port, False)
-    }
-
-
 class LoadBalancer(AWSObject):
     resource_type = "AWS::ElasticLoadBalancing::LoadBalancer"
 
@@ -141,64 +94,5 @@ class LoadBalancer(AWSObject):
         'Scheme': (basestring, False),
         'SecurityGroups': (list, False),
         'Subnets': (list, False),
-        'Tags': (list, False),
-    }
-
-
-class ApplicationListener(AWSObject):
-    resource_type = "AWS::ElasticLoadBalancingV2::Listener"
-
-    props = {
-        'Certificates': ([Certificate], False),
-        'DefaultActions': ([Action], True),
-        'LoadBalancerArn': (basestring, True),
-        'Port': (network_port, True),
-        'Protocol': (basestring, True),
-        'SslPolicy': (basestring, False)
-    }
-
-
-class ApplicationListenerRule(AWSObject):
-    resource_type = "AWS::ElasticLoadBalancingV2::ListenerRule"
-
-    props = {
-        'Actions': ([Action], True),
-        'Conditions': ([Condition], True),
-        'ListenerArn': (basestring, True),
-        'Priority': (integer, True)
-    }
-
-
-class ApplicationTargetGroup(AWSObject):
-    resource_type = "AWS::ElasticLoadBalancingV2::TargetGroup"
-
-    props = {
-        'HealthCheckIntervalSeconds': (integer, False),
-        'HealthCheckPath': (basestring, False),
-        'HealthCheckPort': (network_port, False),
-        'HealthCheckProtocol': (basestring, False),
-        'HealthCheckTimeoutSeconds': (integer, False),
-        'HealthyThresholdCount': (integer, False),
-        'Matcher': (Matcher, False),
-        'Name': (basestring, False),
-        'Port': (network_port, True),
-        'Protocol': (basestring, True),
-        'Tags': (list, False),
-        'TargetGroupAttributes': ([TargetGroupAttribute], False),
-        'Targets': ([TargetDescription], False),
-        'UnhealthyThresholdCount': (integer, False),
-        'VpcId': (basestring, True)
-    }
-
-
-class ApplicationLoadBalancer(AWSObject):
-    resource_type = "AWS::ElasticLoadBalancingV2::LoadBalancer"
-
-    props = {
-        'LoadBalancerAttributes': ([LoadBalancerAttributes], False),
-        'Name': (basestring, False),
-        'Scheme': (basestring, False),
-        'SecurityGroups': (list, False),
-        'Subnets': (list, True),
         'Tags': (list, False),
     }
