+++ troposphere/elasticloadbalancingv2.py
@@ -1,0 +1,114 @@
+# Copyright (c) 2012-2013, Mark Peek <mark@peek.org>
+# All rights reserved.
+#
+# See LICENSE file for full license.
+
+from . import AWSObject, AWSProperty
+from .validators import (
+    network_port, integer)
+
+
+class LoadBalancerAttributes(AWSProperty):
+    props = {
+        'Key': (basestring, False),
+        'Value': (basestring, False)
+    }
+
+
+class Certificate(AWSProperty):
+    props = {
+        'CertificateArn': (basestring, False)
+    }
+
+
+class Action(AWSProperty):
+    props = {
+        'TargetGroupArn': (basestring, True),
+        'Type': (basestring, True)
+    }
+
+
+class Condition(AWSProperty):
+    props = {
+        'Field': (basestring, True),
+        'Values': ([basestring], True)
+    }
+
+
+class Matcher(AWSProperty):
+    props = {
+        'HttpCode': (basestring, False)
+    }
+
+
+class TargetGroupAttribute(AWSProperty):
+    props = {
+        'Key': (basestring, False),
+        'Value': (basestring, False)
+    }
+
+
+class TargetDescription(AWSProperty):
+    props = {
+        'Id': (basestring, True),
+        'Port': (network_port, False)
+    }
+
+
+class Listener(AWSObject):
+    resource_type = "AWS::ElasticLoadBalancingV2::Listener"
+
+    props = {
+        'Certificates': ([Certificate], False),
+        'DefaultActions': ([Action], True),
+        'LoadBalancerArn': (basestring, True),
+        'Port': (network_port, True),
+        'Protocol': (basestring, True),
+        'SslPolicy': (basestring, False)
+    }
+
+
+class ListenerRule(AWSObject):
+    resource_type = "AWS::ElasticLoadBalancingV2::ListenerRule"
+
+    props = {
+        'Actions': ([Action], True),
+        'Conditions': ([Condition], True),
+        'ListenerArn': (basestring, True),
+        'Priority': (integer, True)
+    }
+
+
+class TargetGroup(AWSObject):
+    resource_type = "AWS::ElasticLoadBalancingV2::TargetGroup"
+
+    props = {
+        'HealthCheckIntervalSeconds': (integer, False),
+        'HealthCheckPath': (basestring, False),
+        'HealthCheckPort': (network_port, False),
+        'HealthCheckProtocol': (basestring, False),
+        'HealthCheckTimeoutSeconds': (integer, False),
+        'HealthyThresholdCount': (integer, False),
+        'Matcher': (Matcher, False),
+        'Name': (basestring, False),
+        'Port': (network_port, True),
+        'Protocol': (basestring, True),
+        'Tags': (list, False),
+        'TargetGroupAttributes': ([TargetGroupAttribute], False),
+        'Targets': ([TargetDescription], False),
+        'UnhealthyThresholdCount': (integer, False),
+        'VpcId': (basestring, True)
+    }
+
+
+class LoadBalancer(AWSObject):
+    resource_type = "AWS::ElasticLoadBalancingV2::LoadBalancer"
+
+    props = {
+        'LoadBalancerAttributes': ([LoadBalancerAttributes], False),
+        'Name': (basestring, False),
+        'Scheme': (basestring, False),
+        'SecurityGroups': (list, False),
+        'Subnets': (list, True),
+        'Tags': (list, False),
+    }
